Rebrand core/build logic/relocated classes
