+++ build-logic/settings.gradle.kts
@@ -2,6 +2,6 @@ rootProject.name = "build-logic"
 
 includeBuild("../shared") {
     dependencySubstitution {
-        substitute(module("io.github.juuxel:loom-quiltflower-core")).using(project(":"))
+        substitute(module("io.github.juuxel:vineflower-for-loom-core")).using(project(":"))
     }
 }
