+++ build-logic/src/main/groovy/build-logic.gradle
@@ -1,5 +1,5 @@
-import juuxel.loomquiltflower.plugin.LoomQuiltflowerLogicExtension
-import juuxel.loomquiltflower.plugin.RemapSetup
+import juuxel.vineflowerforloom.plugin.BuildLogicExtension
+import juuxel.vineflowerforloom.plugin.RemapSetup
 
 RemapSetup.setup(project)
-extensions.add("loomQuiltflowerLogic", new LoomQuiltflowerLogicExtension(project))
+extensions.add("buildLogic", new BuildLogicExtension(project))
