+++ build-logic/src/main/java/juuxel/vineflowerforloom/plugin/BuildLogicExtension.java
@@ -1,18 +1,19 @@
-package juuxel.loomquiltflower.plugin;
+package juuxel.vineflowerforloom.plugin;
 
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.Dependency;
 import org.gradle.api.attributes.HasConfigurableAttributes;
 
-public class LoomQuiltflowerLogicExtension {
+public class BuildLogicExtension {
     private final Project project;
 
-    public LoomQuiltflowerLogicExtension(Project project) {
+    public BuildLogicExtension(Project project) {
         this.project = project;
     }
 
-    public Dependency quiltflower() {
-        Dependency dep = project.getDependencies().create("org.quiltmc:quiltflower:" + project.getRootProject().property("quiltflower-version"));
+    public Dependency vineflower() {
+        // TODO: Dep coordinates
+        Dependency dep = project.getDependencies().create("org.quiltmc:quiltflower:" + project.getRootProject().property("vineflower-version"));
         ((HasConfigurableAttributes<?>) dep).attributes(attributes -> attributes.attribute(RemapState.REMAP_STATE_ATTRIBUTE, RemapState.REMAPPED));
         return dep;
     }
