+++ build-logic/src/main/java/juuxel/vineflowerforloom/plugin/RemapSetup.java
@@ -1,4 +1,4 @@
-package juuxel.loomquiltflower.plugin;
+package juuxel.vineflowerforloom.plugin;
 
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.dsl.DependencyHandler;
