+++ settings.gradle.kts
@@ -15,6 +15,6 @@ includeBuild("build-logic")
 
 includeBuild("shared") {
     dependencySubstitution {
-        substitute(module("io.github.juuxel:loom-quiltflower-core")).using(project(":"))
+        substitute(module("io.github.juuxel:vineflower-for-loom-core")).using(project(":"))
     }
 }
