+++ src/loom011/java/juuxel/vineflowerforloom/impl/modern/VineflowerDecompiler.java
@@ -5,9 +5,9 @@ import juuxel.vineflowerforloom.impl.Zips;
 import juuxel.vineflowerforloom.impl.bridge.SimpleLogger;
 import juuxel.vineflowerforloom.impl.bridge.VfResultSaver;
 import juuxel.vineflowerforloom.impl.bridge.VfTinyJavadocProvider;
-import juuxel.loomquiltflower.impl.relocated.quiltflower.main.Fernflower;
-import juuxel.loomquiltflower.impl.relocated.quiltflower.main.extern.IFernflowerPreferences;
-import juuxel.loomquiltflower.impl.relocated.quiltflowerapi.IFabricJavadocProvider;
+import juuxel.vineflowerforloom.impl.relocated.vineflower.main.Fernflower;
+import juuxel.vineflowerforloom.impl.relocated.vineflower.main.extern.IFernflowerPreferences;
+import juuxel.vineflowerforloom.impl.relocated.vineflowerapi.IFabricJavadocProvider;
 import net.fabricmc.loom.api.decompilers.DecompilationMetadata;
 import net.fabricmc.loom.api.decompilers.LoomDecompiler;
 import net.fabricmc.loom.util.IOStringConsumer;
