+++ src/main/java/juuxel/vineflowerforloom/impl/bridge/SimpleLogger.java
@@ -1,6 +1,6 @@
 package juuxel.vineflowerforloom.impl.bridge;
 
-import juuxel.loomquiltflower.impl.relocated.quiltflower.main.extern.IFernflowerLogger;
+import juuxel.vineflowerforloom.impl.relocated.vineflower.main.extern.IFernflowerLogger;
 
 import java.io.PrintWriter;
